content/canvas/test/test_2d.drawImage.negativesource.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.drawImage.negativesource</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
function isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    ok(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.drawImage(document.getElementById('ggrr-256x256.png'), 100, 78, -100, 50, 0, 0, 50, 50);
ctx.drawImage(document.getElementById('ggrr-256x256.png'), 100, 128, -100, -50, 50, 0, 50, 50);
isPixel(ctx, 1,1, 0,255,0,255, "1,1", "0,255,0,255", 2);
isPixel(ctx, 1,48, 0,255,0,255, "1,48", "0,255,0,255", 2);
isPixel(ctx, 98,1, 0,255,0,255, "98,1", "0,255,0,255", 2);
isPixel(ctx, 98,48, 0,255,0,255, "98,48", "0,255,0,255", 2);
isPixel(ctx, 48,1, 0,255,0,255, "48,1", "0,255,0,255", 2);
isPixel(ctx, 48,48, 0,255,0,255, "48,48", "0,255,0,255", 2);
isPixel(ctx, 51,1, 0,255,0,255, "51,1", "0,255,0,255", 2);
isPixel(ctx, 51,48, 0,255,0,255, "51,48", "0,255,0,255", 2);
isPixel(ctx, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
isPixel(ctx, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
<img src="image_ggrr-256x256.png" id="ggrr-256x256.png" class="resource">
